Add instructor endpoint for FCEs
