+++ backend/src/app.js
@@ -1,26 +1,26 @@
 import express from "express";
 import mongoose from "mongoose";
 import cors from "cors";
-import { KeyStore, generateSigningRequestHandler } from "passlink-server";
+import { generateSigningRequestHandler, KeyStore } from "passlink-server";
 import { isUser } from "./controllers/user.js";
 import {
-	getCourseByID,
-	getCourses,
-	getAllCourses,
-	getFilteredCourses,
+  getAllCourses,
+  getCourseByID,
+  getCourses,
+  getFilteredCourses,
 } from "./controllers/courses.js";
-import { getFCEsByID } from "./controllers/fces.js";
+import { getFCEs } from "./controllers/fces.js";
 
 KeyStore.readKey(process.env.LOGIN_API_KEY);
 
 const app = express();
 const signingRequestHandler = generateSigningRequestHandler(
-	{
-		restrictDomain: true,
-		applicationId: process.env.LOGIN_API_ID,
-	},
-	KeyStore.getSecretKey(),
-	true
+  {
+    restrictDomain: true,
+    applicationId: process.env.LOGIN_API_ID,
+  },
+  KeyStore.getSecretKey(),
+  true
 );
 
 const port = process.env.PORT || 3000;
@@ -35,7 +35,7 @@ app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
 
 app.get("/signingrequest", (req, res) => {
-	signingRequestHandler(req, res);
+  signingRequestHandler(req, res);
 });
 
 app.route("/course/:courseID").get(getCourseByID);
@@ -45,8 +45,8 @@ app.route("/courses/all").get(getAllCourses);
 app.route("/courses/search/").get(getFilteredCourses);
 app.route("/courses/search/").post(isUser, getFilteredCourses);
 
-app.route("/fces").post(isUser, getFCEsByID);
+app.route("/fces").post(isUser, getFCEs);
 
 app.listen(port, () =>
-	console.log(`Course Tool backend listening on port ${port}.`)
+  console.log(`Course Tool backend listening on port ${port}.`)
 );
