+++ backend/src/controllers/courses.js
@@ -1,106 +1,106 @@
-import { standardizeID, singleToArray } from "../util.js";
+import { singleToArray, standardizeID } from "../util.js";
 import { Course, Schedule } from "../models/course.js";
 
 const projection = { _id: false, __v: false };
 const MAX_LIMIT = 10;
 
 export const getCourseByID = (req, res) => {
-	const id = standardizeID(req.params.courseID);
+  const id = standardizeID(req.params.courseID);
 
-	const options = {};
-	if ("schedules" in req.query && req.query.schedules)
-		options.populate = [{ path: "schedules", model: Schedule }];
+  const options = {};
+  if ("schedules" in req.query && req.query.schedules)
+    options.populate = [{ path: "schedules", model: Schedule }];
 
-	Course.findOne({ courseID: id }, projection, options)
-		.then((result) => res.json(result))
-		.catch((err) => res.status(500).send(err));
+  Course.findOne({ courseID: id }, projection, options)
+    .then((result) => res.json(result))
+    .catch((err) => res.status(500).send(err));
 };
 
 export const getCourses = (req, res) => {
-	const courseIDs = singleToArray(req.query.courseID).map(standardizeID);
+  const courseIDs = singleToArray(req.query.courseID).map(standardizeID);
 
-	const options = { populate: [] };
-	if ("schedulesAvailable" in req.query && req.query.schedulesAvailable) {
-		options.populate.push({
-			path: "schedules",
-			model: Schedule,
-			select: "year semester session _id",
-		});
-	}
+  const options = { populate: [] };
+  if ("schedulesAvailable" in req.query && req.query.schedulesAvailable) {
+    options.populate.push({
+      path: "schedules",
+      model: Schedule,
+      select: "year semester session _id",
+    });
+  }
 
-	Course.find({ courseID: { $in: courseIDs } }, projection, options)
-		.then((result) => res.json(result))
-		.catch((err) => res.status(500).send(err));
+  Course.find({ courseID: { $in: courseIDs } }, projection, options)
+    .then((result) => res.json(result))
+    .catch((err) => res.status(500).send(err));
 };
 
 export const getFilteredCourses = (req, res) => {
-	const matchQuery = {};
-	const options = {
-		projection: {...projection},
-		limit: MAX_LIMIT,
-		populate: [],
-	};
+  const matchQuery = {};
+  const options = {
+    projection: { ...projection },
+    limit: MAX_LIMIT,
+    populate: [],
+  };
 
-	if ("department" in req.query)
-		matchQuery["department"] = { $in: singleToArray(req.query.department) };
+  if ("department" in req.query)
+    matchQuery["department"] = { $in: singleToArray(req.query.department) };
 
-	if ("keywords" in req.query) {
-		matchQuery["$text"] = { $search: req.query.keywords };
-		options.projection.score = { $meta: "textScore" };
-	}
+  if ("keywords" in req.query) {
+    matchQuery["$text"] = { $search: req.query.keywords };
+    options.projection.score = { $meta: "textScore" };
+  }
 
-	if ("keywords" in req.query) options.sort = { score: { $meta: "textScore" } };
+  if ("keywords" in req.query) options.sort = { score: { $meta: "textScore" } };
 
-	if ("page" in req.query) options.page = req.query.page;
+  if ("page" in req.query) options.page = req.query.page;
 
-	if ("schedules" in req.query && req.query.schedules)
-		options.populate.push({
-			path: "schedules",
-			model: Schedule,
-			select: "-_id",
-		});
+  if ("schedules" in req.query && req.query.schedules)
+    options.populate.push({
+      path: "schedules",
+      model: Schedule,
+      select: "-_id",
+    });
 
-	if ("schedulesAvailable" in req.query && req.query.schedulesAvailable)
-		options.populate.push({
-			path: "schedules",
-			model: Schedule,
-			select: "year semester session -_id",
-		});
+  if ("schedulesAvailable" in req.query && req.query.schedulesAvailable)
+    options.populate.push({
+      path: "schedules",
+      model: Schedule,
+      select: "year semester session -_id",
+    });
 
-	if (req.method === "POST") {
-		if ("fces" in req.query && req.query.fces) {
-			options.populate.push({
-				path: "fces",
-				model: FCE,
-				select: "-_id",
-			});
-		}
-	}
+  if (req.method === "POST") {
+    if ("fces" in req.query && req.query.fces) {
+      options.populate.push({
+        path: "fces",
+        model: FCE,
+        select: "-_id",
+      });
+    }
+  }
 
-	Course.paginate(matchQuery, options)
-		.then((result) => res.json(result))
-		.catch((err) => res.status(500).send(err));
+  Course.paginate(matchQuery, options)
+    .then((result) => res.json(result))
+    .catch((err) => res.status(500).send(err));
 };
 
 // TODO: use a better caching system
 const allCoursesEntry = {
-	allCourses: [],
-	lastCached: null,
+  allCourses: [],
+  lastCached: null,
 };
 
 export const getAllCourses = (req, res) => {
-	if (
-		allCoursesEntry.lastCached === null ||
-		new Date() - allCoursesEntry.lastCached > 1000 * 60 * 60 * 24
-	) {
-		Course.find({}, "courseID name -_id")
-			.then((result) => {
-				allCoursesEntry.lastCached = new Date();
-				allCoursesEntry.allCourses = result;
-				res.json(result);
-			})
-			.catch((err) => res.status(500).send(err));
-	} else {
-		res.json(allCoursesEntry.allCourses);
-	}
+  if (
+    allCoursesEntry.lastCached === null ||
+    new Date() - allCoursesEntry.lastCached > 1000 * 60 * 60 * 24
+  ) {
+    Course.find({}, "courseID name -_id")
+      .then((result) => {
+        allCoursesEntry.lastCached = new Date();
+        allCoursesEntry.allCourses = result;
+        res.json(result);
+      })
+      .catch((err) => res.status(500).send(err));
+  } else {
+    res.json(allCoursesEntry.allCourses);
+  }
 };
