+++ backend/src/models/professor.js
@@ -1,20 +1,20 @@
 import mongoose from "mongoose";
 
 export const profSchema = new mongoose.Schema({
-	name: String,
-	andrewID: String,
-	departments: [String],
-	courses: [
-		{
-			courseID: String,
-			yearHrsPerWeek: Number,
-			summerHrsPerWeek: Number,
-			teachingRating: Number,
-			overallRating: Number,
-			semesters: [{ year: String, semester: String }],
-		},
-	],
-	respectForStudents: Number,
-	interestInLearning: Number,
-	providesFeedback: Number,
+  name: String,
+  andrewID: String,
+  departments: [String],
+  courses: [
+    {
+      courseID: String,
+      yearHrsPerWeek: Number,
+      summerHrsPerWeek: Number,
+      teachingRating: Number,
+      overallRating: Number,
+      semesters: [{ year: String, semester: String }],
+    },
+  ],
+  respectForStudents: Number,
+  interestInLearning: Number,
+  providesFeedback: Number,
 });
